requirements.txt
kafka-python==2.0.2
pandas==2.1.0
numpy==1.25.0
xgboost==2.0.0
scikit-learn==1.3.0
psycopg2-binary==2.9.9
fastapi==0.104.0
uvicorn==0.24.0
python-dotenv==1.0.0
